+++ linux_voice_assistant/__main__.py
@@ -18,7 +18,7 @@ from getmac import get_mac_address  # type: ignore
 from pymicro_wakeword import MicroWakeWord, MicroWakeWordFeatures
 from pyopen_wakeword import OpenWakeWord, OpenWakeWordFeatures
 
-from .models import Preferences, ServerState, WakeWordType
+from .models import Preferences, ServerState, WakeWordType, initial_stop_word_threshold
 from .mpv_player import MpvMediaPlayer
 from .peripheral_api import LVAEvent, PeripheralAPIServer
 from .satellite import VoiceSatelliteProtocol
@@ -380,8 +380,8 @@ async def main() -> None:
     preferences.volume = initial_volume
 
     # Load stop word sensitivity from preferences on startup, and ensure it's between 0.0 and 1.0
-    initial_stop_word_threshold = _initial_stop_word_threshold(preferences.stop_word_sensitivity)
-    preferences.stop_word_sensitivity = initial_stop_word_threshold
+    initial_threshold = initial_stop_word_threshold(preferences.stop_word_sensitivity)
+    preferences.stop_word_sensitivity = initial_threshold
 
     if args.enable_thinking_sound:
         preferences.thinking_sound = 1
@@ -445,7 +445,7 @@ async def main() -> None:
         output_only=args.output_only,
         download_dir=args.download_dir,
         volume=initial_volume,
-        stop_word_threshold=initial_stop_word_threshold,
+        stop_word_threshold=initial_threshold,
         mic_volume=preferences.mic_volume,
         mic_auto_gain=preferences.mic_auto_gain,
         mic_noise_suppression=preferences.mic_noise_suppression,
@@ -593,18 +593,6 @@ async def main() -> None:
             await peripheral_api.stop()
 
     _LOGGER.debug("Server stopped")
-
-
-# -----------------------------------------------------------------------------
-def _initial_stop_word_threshold(saved_sensitivity: Optional[float]) -> float:
-    """
-    Resolve the stop word probability cutoff to start from, clamped to 0.0-1.0.
-    :param saved_sensitivity: Value persisted in preferences, or None if it has never been set.
-    """
-    if saved_sensitivity is None:
-        return ServerState.stop_word_threshold
-
-    return max(0.0, min(1.0, float(saved_sensitivity)))
 
 
 # -----------------------------------------------------------------------------
